Revert "ran Analyze to clean up code (no logic changes)"
This reverts commit 2174ce76dadc195bb15080910965bd6c5f69f8a1.

+++ FtcRobotController/src/main/java/com/qualcomm/hardware/ModernRoboticsUsbServoController.java
@@ -51,11 +51,11 @@ public class ModernRoboticsUsbServoController extends ModernRoboticsUsbDevice im
     }
 
     public void pwmEnable() {
-        write(ADDRESS_PWM, PWM_ENABLE);
+        write((int) ADDRESS_PWM, (byte) PWM_ENABLE);
     }
 
     public void pwmDisable() {
-        write(ADDRESS_PWM, PWM_DISABLE);
+        write((int) ADDRESS_PWM, (byte) PWM_DISABLE);
     }
 
     public PwmStatus getPwmStatus() {
@@ -78,8 +78,8 @@ public class ModernRoboticsUsbServoController extends ModernRoboticsUsbDevice im
     }
 
     private void m72a(int i) {
-        if ((i < 1) || (i > ADDRESS_CHANNEL_MAP.length)) {
-            throw new IllegalArgumentException(String.format("Channel %d is invalid; valid channels are 1..%d", new Object[]{i, MAX_SERVOS}));
+        if (i < 1 || i > ADDRESS_CHANNEL_MAP.length) {
+            throw new IllegalArgumentException(String.format("Channel %d is invalid; valid channels are 1..%d", new Object[]{Integer.valueOf(i), Integer.valueOf(MAX_SERVOS)}));
         }
     }
 }
